Moved new_handlers, new_jobs to handlers, jobs
