+++ custodian/qchem/handlers.py
@@ -2,551 +2,223 @@
 
 from __future__ import unicode_literals, division
 
-import shutil
-import time
+# This module implements new error handlers for QChem runs.
 
-"""
-This module implements error handlers for QChem runs. Currently tested only
-for B3LYP DFT jobs.
-"""
-
-import copy
-import glob
-import json
-import logging
 import os
-import re
-import tarfile
-from pymatgen.core.structure import Molecule
-from pymatgen.io.qchem import QcOutput, QcInput, QcTask
+from pymatgen.io.qchem.inputs import QCInput
+from pymatgen.io.qchem.outputs import QCOutput
 from custodian.custodian import ErrorHandler
+from custodian.utils import backup
 
-__author__ = "Xiaohui Qu"
+__author__ = "Samuel Blau, Brandon Wood, Shyam Dwaraknath"
+__copyright__ = "Copyright 2018, The Materials Project"
 __version__ = "0.1"
-__maintainer__ = "Xiaohui Qu"
-__email__ = "xhqu1981@gmail.com"
+__maintainer__ = "Samuel Blau"
+__email__ = "samblau1@gmail.com"
 __status__ = "Alpha"
-__date__ = "12/04/13"
+__date__ = "3/26/18"
+__credits__ = "Xiaohui Qu"
 
 
 class QChemErrorHandler(ErrorHandler):
     """
-    Error handler for QChem Jobs. Currently tested only for B3LYP DFT jobs
-    generated by pymatgen.
+    Master QChemErrorHandler class that handles a number of common errors
+    that occur during QChem runs.
     """
-    def __init__(self, input_file="mol.qcinp", output_file="mol.qcout",
-                 ex_backup_list=(), rca_gdm_thresh=1.0E-3,
-                 scf_max_cycles=200, geom_max_cycles=200, qchem_job=None):
+
+    is_monitor = False
+
+    def __init__(self,
+                 input_file="mol.qin",
+                 output_file="mol.qout",
+                 scf_max_cycles=200,
+                 geom_max_cycles=200):
         """
         Initializes the error handler from a set of input and output files.
 
         Args:
             input_file (str): Name of the QChem input file.
             output_file (str): Name of the QChem output file.
-            ex_backup_list ([str]): List of the files to backup in addition
-                to input and output file.
+            scf_max_cycles (int): The max iterations to set to fix SCF failure.
+            geom_max_cycles (int): The max iterations to set to fix geometry
+                optimization failure.
+        """
+        self.input_file = input_file
+        self.output_file = output_file
+        self.scf_max_cycles = scf_max_cycles
+        self.geom_max_cycles = geom_max_cycles
+        # self.qcinp = QCInput.from_file(self.input_file)
+        self.outdata = None
+        self.errors = []
+
+    def check(self):
+        # Checks output file for errors.
+        self.outdata = QCOutput(self.output_file).data
+        self.errors = self.outdata.get("errors")
+        return len(self.errors) > 0
+
+    def correct(self):
+        backup({self.input_file, self.output_file})
+        self.qcinp = QCInput.from_file(self.input_file)
+        actions = []
+
+        if "SCF_failed_to_converge" in self.errors:
+            # Check number of SCF cycles. If not set or less than scf_max_cycles,
+            # increase to that value and rerun. If already set, check if
+            # scf_algorithm is unset or set to DIIS, in which case set to RCA-DIIS.
+            # Otherwise, tell user to call SCF error handler and do nothing.
+            if self.qcinp.rem.get("max_scf_cycles") != str(
+                    self.scf_max_cycles):
+                self.qcinp.rem["max_scf_cycles"] = self.scf_max_cycles
+                actions.append({"max_scf_cycles": self.scf_max_cycles})
+            elif self.qcinp.rem.get("scf_algorithm", "diis").lower() == "diis":
+                self.qcinp.rem["scf_algorithm"] = "rca_diis"
+                actions.append({"scf_algorithm": "rca_diis"})
+                if self.qcinp.rem.get("gen_scfman"):
+                    self.qcinp.rem["gen_scfman"] = False
+                    actions.append({"gen_scfman": False})
+            else:
+                print(
+                    "More advanced changes may impact the SCF result. Use the SCF error handler"
+                )
+
+        elif "out_of_opt_cycles" in self.errors:
+            # Check number of opt cycles. If less than geom_max_cycles, increase
+            # to that value, set last geom as new starting geom and rerun.
+            if self.qcinp.rem.get(
+                    "geom_opt_max_cycles") != self.geom_max_cycles:
+                self.qcinp.rem["geom_opt_max_cycles"] = self.geom_max_cycles
+                actions.append({"geom_max_cycles:": self.scf_max_cycles})
+                if len(self.outdata.get("energy_trajectory")) > 1:
+                    if self.qcinp.molecule.spin_multiplicity != self.outdata.get(
+                            "molecule_from_last_geometry").spin_multiplicity:
+                        raise AssertionError('Multiplicities should match!')
+                    if self.qcinp.molecule.charge != self.outdata.get(
+                            "molecule_from_last_geometry").charge:
+                        raise AssertionError('Charges should match!')
+                    self.qcinp.molecule = self.outdata.get(
+                        "molecule_from_last_geometry")
+                    actions.append({"molecule": "molecule_from_last_geometry"})
+            else:
+                print(
+                    "How do I get the geometry optimization converged when already at the maximum number of cycles?"
+                )
+
+        elif "unable_to_determine_lamda" in self.errors:
+            # Set last geom as new starting geom and rerun. If no opt cycles,
+            # use diff SCF strat? Diff initial guess? Change basis?
+            if len(self.outdata.get("energy_trajectory")) > 1:
+                if self.qcinp.molecule.spin_multiplicity != self.outdata.get(
+                        "molecule_from_last_geometry").spin_multiplicity:
+                    raise AssertionError('Multiplicities should match!')
+                if self.qcinp.molecule.charge != self.outdata.get(
+                        "molecule_from_last_geometry").charge:
+                    raise AssertionError('Charges should match!')
+                self.qcinp.molecule = self.outdata.get(
+                    "molecule_from_last_geometry")
+                actions.append({"molecule": "molecule_from_last_geometry"})
+            elif self.qcinp.rem.get("scf_algorithm", "diis").lower() == "diis":
+                self.qcinp.rem["scf_algorithm"] = "rca_diis"
+                actions.append({"scf_algorithm": "rca_diis"})
+                if self.qcinp.rem.get("gen_scfman"):
+                    self.qcinp.rem["gen_scfman"] = False
+                    actions.append({"gen_scfman": False})
+            else:
+                print(
+                    "Use a different initial guess? Perhaps a different basis?"
+                )
+
+        elif "linear_dependent_basis" in self.errors:
+            # DIIS -> RCA_DIIS. If already RCA_DIIS, change basis?
+            if self.qcinp.rem.get("scf_algorithm", "diis").lower() == "diis":
+                self.qcinp.rem["scf_algorithm"] = "rca_diis"
+                actions.append({"scf_algorithm": "rca_diis"})
+                if self.qcinp.rem.get("gen_scfman"):
+                    self.qcinp.rem["gen_scfman"] = False
+                    actions.append({"gen_scfman": False})
+            else:
+                print("Perhaps use a better basis?")
+
+        elif "failed_to_transform_coords" in self.errors:
+            # Check for symmetry flag in rem. If not False, set to False and rerun.
+            # If already False, increase threshold?
+            if not self.qcinp.rem.get("sym_ignore") or self.qcinp.rem.get(
+                    "symmetry"):
+                self.qcinp.rem["sym_ignore"] = True
+                self.qcinp.rem["symmetry"] = False
+                actions.append({"sym_ignore": True})
+                actions.append({"symmetry": False})
+            else:
+                print("Perhaps increase the threshold?")
+
+        elif "input_file_error" in self.errors:
+            print(
+                "Something is wrong with the input file. Examine error message by hand."
+            )
+            return {"errors": self.errors, "actions": None}
+
+        elif "failed_to_read_input" in self.errors:
+            # Almost certainly just a temporary problem that will not be encountered again. Rerun job as-is.
+            actions.append({"rerun job as-is"})
+
+        elif "IO_error" in self.errors:
+            # Almost certainly just a temporary problem that will not be encountered again. Rerun job as-is.
+            actions.append({"rerun job as-is"})
+
+        elif "unknown_error" in self.errors:
+            print("Examine error message by hand.")
+            return {"errors": self.errors, "actions": None}
+
+        else:
+            # You should never get here. If correct is being called then errors should have at least one entry,
+            # in which case it should have been caught by the if/elifs above.
+            print(
+                "If you get this message, something has gone terribly wrong!")
+            return {"errors": self.errors, "actions": None}
+
+        os.rename(self.input_file, self.input_file + ".last")
+        self.qcinp.write_file(self.input_file)
+        return {"errors": self.errors, "actions": actions}
+
+
+class QChemSCFErrorHandler(ErrorHandler):
+    """
+    QChem ErrorHandler class that addresses SCF non-convergence.
+    """
+
+    is_monitor = False
+
+    def __init__(self,
+                 input_file="mol.qin",
+                 output_file="mol.qout",
+                 rca_gdm_thresh=1.0E-3,
+                 scf_max_cycles=200):
+        """
+        Initializes the error handler from a set of input and output files.
+
+        Args:
+            input_file (str): Name of the QChem input file.
+            output_file (str): Name of the QChem output file.
             rca_gdm_thresh (float): The threshold for the prior scf algorithm.
                 If last deltaE is larger than the threshold try RCA_DIIS
                 first, else, try DIIS_GDM first.
             scf_max_cycles (int): The max iterations to set to fix SCF failure.
-            geom_max_cycles (int): The max iterations to set to fix geometry
-                optimization failure.
-            qchem_job (QchemJob): the managing object to run qchem.
         """
         self.input_file = input_file
         self.output_file = output_file
-        self.ex_backup_list = ex_backup_list
-        self.rca_gdm_thresh = rca_gdm_thresh
         self.scf_max_cycles = scf_max_cycles
         self.geom_max_cycles = geom_max_cycles
+        self.qcinp = QCInput.from_file(self.input_file)
         self.outdata = None
-        self.qcinp = None
-        self.error_step_id = None
         self.errors = None
-        self.fix_step = None
         self.qchem_job = qchem_job
 
     def check(self):
         # Checks output file for errors.
-        self.outdata = QcOutput(self.output_file).data
-        self.qcinp = QcInput.from_file(self.input_file)
-        self.error_step_id = None
-        self.errors = None
-        self.fix_step = None
-        for i, od in enumerate(self.outdata):
-            if od["has_error"]:
-                self.error_step_id = i
-                self.fix_step = self.qcinp.jobs[i]
-                self.errors = sorted(list(set(od["errors"])))
-                return True
-        return False
+        self.outdata = QCOutput(self.output_file).data
+        self.errors = self.outdata.get("errors")
+        return len(self.errors) > 0
 
     def correct(self):
-        self.backup()
-        actions = []
-        error_rankings = ("pcm_solvent deprecated",
-                          "autoz error",
-                          "No input text",
-                          "Killed",
-                          "Insufficient static memory",
-                          "Not Enough Total Memory",
-                          "NAN values",
-                          "Bad SCF convergence",
-                          "Geometry optimization failed",
-                          "Freq Job Too Small",
-                          "Exit Code 134",
-                          "Molecular charge is not found",
-                          "Molecular spin multipilicity is not found"
-                          )
-        e = self.errors[0]
-        for prio_error in error_rankings:
-            if prio_error in self.errors:
-                e = prio_error
-                break
-
-        if e == "autoz error":
-            if "sym_ignore" not in self.fix_step.params["rem"]:
-                self.fix_step.disable_symmetry()
-                actions.append("disable symmetry")
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Bad SCF convergence":
-            act = self.fix_scf()
-            if act:
-                actions.append(act)
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Geometry optimization failed":
-            act = self.fix_geom_opt()
-            if act:
-                actions.append(act)
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "NAN values":
-            if "xc_grid" not in self.fix_step.params["rem"]:
-                self.fix_step.set_dft_grid(128, 302)
-                actions.append("use tighter grid")
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "No input text":
-            if "sym_ignore" not in self.fix_step.params["rem"]:
-                self.fix_step.disable_symmetry()
-                actions.append("disable symmetry")
-            else:
-                # This indicates something strange occured on the
-                # compute node. Wait for 30 minutes, such that it
-                # won't run too fast to make all the jobs fail
-                if "PBS_JOBID" in os.environ and ("edique" in os.environ["PBS_JOBID"]
-                                                  or "hopque" in os.environ["PBS_JOBID"]):
-                    time.sleep(30.0 * 60.0)
-                return {"errors": self.errors, "actions": None}
-        elif e == "Freq Job Too Small":
-            natoms = len(self.fix_step.mol)
-            if "cpscf_nseg" not in self.fix_step.params["rem"] or \
-                            self.fix_step.params["rem"]["cpscf_nseg"] != natoms:
-                self.fix_step.params["rem"]["cpscf_nseg"] = natoms
-                actions.append("use {} segment in CPSCF".format(natoms))
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "pcm_solvent deprecated":
-            solvent_params = self.fix_step.params.pop("pcm_solvent", None)
-            if solvent_params is not None:
-                self.fix_step.params["solvent"] = solvent_params
-                actions.append("use keyword solvent instead")
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Exit Code 134":
-            act = self.fix_error_code_134()
-            if act:
-                actions.append(act)
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Killed":
-            act = self.fix_error_killed()
-            if act:
-                actions.append(act)
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Insufficient static memory":
-            act = self.fix_insufficient_static_memory()
-            if act:
-                actions.append(act)
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Not Enough Total Memory":
-            act = self.fix_not_enough_total_memory()
-            if act:
-                actions.append(act)
-            else:
-                return {"errors": self.errors, "actions": None}
-        elif e == "Molecular charge is not found":
-            return {"errors": self.errors, "actions": None}
-        elif e == "Molecular spin multipilicity is not found":
-            return {"errors": self.errors, "actions": None}
-        else:
-            return {"errors": self.errors, "actions": None}
-        self.qcinp.write_file(self.input_file)
-        return {"errors": self.errors, "actions": actions}
-
-    def fix_not_enough_total_memory(self):
-        if self.fix_step.params['rem']["jobtype"] == "freq":
-            ncpu = 1
-            if "-np" in self.qchem_job.current_command:
-                cmd = self.qchem_job.current_command
-                ncpu = int(cmd[cmd.index("-np") + 1])
-            natoms = len(self.qcinp.jobs[0].mol)
-            times_ncpu_full = int(natoms/ncpu)
-            nsegment_full = ncpu * times_ncpu_full
-            times_ncpu_half = int(natoms/(ncpu/2))
-            nsegment_half = int((ncpu/2) * times_ncpu_half)
-            if "cpscf_nseg" not in self.fix_step.params["rem"]:
-                self.fix_step.params["rem"]["cpscf_nseg"] = nsegment_full
-                return "Use {} CPSCF segments".format(nsegment_full)
-            elif self.fix_step.params["rem"]["cpscf_nseg"] < nsegment_half:
-                self.qchem_job.select_command("half_cpus", self.qcinp)
-                self.fix_step.params["rem"]["cpscf_nseg"] = nsegment_half
-                return "Use half CPUs and {} CPSCF segments".format(nsegment_half)
-            return None
-        elif not self.qchem_job.is_openmp_compatible(self.qcinp):
-            if self.qchem_job.current_command_name != "half_cpus":
-                self.qchem_job.select_command("half_cpus", self.qcinp)
-                return "half_cpus"
-            else:
-                return None
-
-    def fix_error_code_134(self):
-
-        if "thresh" not in self.fix_step.params["rem"]:
-            self.fix_step.set_integral_threshold(thresh=12)
-            return "use tight integral threshold"
-        elif not (self.qchem_job.is_openmp_compatible(self.qcinp) and
-                      self.qchem_job.command_available("openmp")):
-            if self.qchem_job.current_command_name != "half_cpus":
-                self.qchem_job.select_command("half_cpus", self.qcinp)
-                return "half_cpus"
-            else:
-                if self.fix_step.params['rem']["jobtype"] == "freq":
-                    act = self.fix_not_enough_total_memory()
-                    return act
-                return None
-        elif self.qchem_job.current_command_name != "openmp":
-            self.qchem_job.select_command("openmp", self.qcinp)
-            return "openmp"
-        else:
-            if self.fix_step.params['rem']["jobtype"] == "freq":
-                act = self.fix_not_enough_total_memory()
-                return act
-            return None
-
-    def fix_insufficient_static_memory(self):
-        if not (self.qchem_job.is_openmp_compatible(self.qcinp)
-                    and self.qchem_job.command_available("openmp")):
-            if self.qchem_job.current_command_name != "half_cpus":
-                self.qchem_job.select_command("half_cpus", self.qcinp)
-                return "half_cpus"
-            elif not self.qchem_job.large_static_mem:
-                self.qchem_job.large_static_mem = True
-                # noinspection PyProtectedMember
-                self.qchem_job._set_qchem_memory(self.qcinp)
-                return "Increase Static Memory"
-            else:
-                return None
-        elif self.qchem_job.current_command_name != "openmp":
-            self.qchem_job.select_command("openmp", self.qcinp)
-            return "Use OpenMP"
-        elif not self.qchem_job.large_static_mem:
-            self.qchem_job.large_static_mem = True
-            # noinspection PyProtectedMember
-            self.qchem_job._set_qchem_memory(self.qcinp)
-            return "Increase Static Memory"
-        else:
-            return None
-
-    def fix_error_killed(self):
-        if not (self.qchem_job.is_openmp_compatible(self.qcinp)
-                    and self.qchem_job.command_available("openmp")):
-            if self.qchem_job.current_command_name != "half_cpus":
-                self.qchem_job.select_command("half_cpus", self.qcinp)
-                return "half_cpus"
-            else:
-                return None
-        elif self.qchem_job.current_command_name != "openmp":
-            self.qchem_job.select_command("openmp", self.qcinp)
-            return "Use OpenMP"
-        else:
-            return None
-
-    def fix_scf(self):
-        comments = self.fix_step.params.get("comment", "")
-        scf_pattern = re.compile(r"<SCF Fix Strategy>(.*)</SCF Fix "
-                                 r"Strategy>", flags=re.DOTALL)
-        old_strategy_text = re.findall(scf_pattern, comments)
-
-        if len(old_strategy_text) > 0:
-            old_strategy_text = old_strategy_text[0]
-        od = self.outdata[self.error_step_id]
-
-        if "Negative Eigen" in self.errors:
-            if "thresh" not in self.fix_step.params["rem"]:
-                self.fix_step.set_integral_threshold(thresh=12)
-                return "use tight integral threshold"
-            elif int(self.fix_step.params["rem"]["thresh"]) < 14:
-                self.fix_step.set_integral_threshold(thresh=14)
-                return "use even tighter integral threshold"
-
-        if len(od["scf_iteration_energies"]) == 0 \
-                or len(od["scf_iteration_energies"][-1]) <= 10:
-            if 'Exit Code 134' in self.errors:
-                # immature termination of SCF
-                return self.fix_error_code_134()
-            else:
-                return None
-
-        if od["jobtype"] in ["opt", "ts", "aimd"] \
-                and len(od["molecules"]) >= 2:
-            strategy = "reset"
-        elif len(old_strategy_text) > 0:
-            strategy = json.loads(old_strategy_text)
-            strategy["current_method_id"] += 1
-        else:
-            strategy = dict()
-            scf_iters = od["scf_iteration_energies"][-1]
-            if scf_iters[-1][1] >= self.rca_gdm_thresh:
-                strategy["methods"] = ["increase_iter", "rca_diis", "gwh",
-                                       "gdm", "rca", "core+rca", "fon"]
-                strategy["current_method_id"] = 0
-            else:
-                strategy["methods"] = ["increase_iter", "diis_gdm", "gwh",
-                                       "rca", "gdm", "core+gdm", "fon"]
-                strategy["current_method_id"] = 0
-            strategy["version"] = 2.0
-
-        # noinspection PyTypeChecker
-        if strategy == "reset":
-            self.fix_step.set_scf_algorithm_and_iterations(
-                algorithm="diis", iterations=self.scf_max_cycles)
-            if self.error_step_id > 0:
-                self.set_scf_initial_guess("read")
-            else:
-                self.set_scf_initial_guess("sad")
-            if od["jobtype"] in ["opt", "ts"]:
-                self.set_last_input_geom(od["molecules"][-1])
-            else:
-                assert od["jobtype"] == "aimd"
-                from pymatgen.io.qchem import QcNucVeloc
-                from pymatgen.io.xyz import XYZ
-                scr_dir = od["scratch_dir"]
-                qcnv_filepath = os.path.join(scr_dir, "AIMD", "NucVeloc")
-                qc_md_view_filepath = os.path.join(scr_dir, "AIMD", "View.xyz")
-                qcnv = QcNucVeloc(qcnv_filepath)
-                qc_md_view = XYZ.from_file(qc_md_view_filepath)
-                assert len(qcnv.velocities) == len(qc_md_view.all_molecules)
-                aimd_steps = self.fix_step.params["rem"]["aimd_steps"]
-                elapsed_steps = len(qc_md_view.all_molecules)
-                remaining_steps = aimd_steps - elapsed_steps + 1
-                self.fix_step.params["rem"]["aimd_steps"] = remaining_steps
-                self.set_last_input_geom(qc_md_view.molecule)
-                self.fix_step.set_velocities(qcnv.velocities[-1])
-                self.fix_step.params["rem"].pop("aimd_init_veloc", None)
-                traj_num = max([0] + [int(f.split(".")[1])
-                                       for f in glob.glob("traj_View.*.xyz")])
-                dest_view_filename = "traj_View.{}.xyz".format(traj_num + 1)
-                dest_nv_filename = "traj_NucVeloc.{}.txt".format(traj_num + 1)
-                logging.info("Backing up trajectory files to {} and {}."
-                             .format(dest_view_filename, dest_nv_filename))
-                shutil.copy(qc_md_view_filepath, dest_view_filename)
-                shutil.copy(qcnv_filepath, dest_nv_filename)
-            if len(old_strategy_text) > 0:
-                comments = scf_pattern.sub("", comments)
-                self.fix_step.params["comment"] = comments
-                if len(comments.strip()) == 0:
-                    self.fix_step.params.pop("comment")
-            return "reset"
-        elif strategy["current_method_id"] > len(strategy["methods"])-1:
-            return None
-        else:
-            # noinspection PyTypeChecker
-            method = strategy["methods"][strategy["current_method_id"]]
-            if method == "increase_iter":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="diis", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("sad")
-            elif method == "rca_diis":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="rca_diis", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("sad")
-            elif method == "gwh":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="diis", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("gwh")
-            elif method == "gdm":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="gdm", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("sad")
-            elif method == "rca":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="rca", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("sad")
-            elif method == "core+rca":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="rca", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("core")
-            elif method == "diis_gdm":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="diis_gdm", iterations=self.scf_max_cycles)
-                self.fix_step.set_scf_initial_guess("sad")
-            elif method == "core+gdm":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="gdm", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("core")
-            elif method == "fon":
-                self.fix_step.set_scf_algorithm_and_iterations(
-                    algorithm="diis", iterations=self.scf_max_cycles)
-                self.set_scf_initial_guess("sad")
-                natoms = len(od["molecules"][-1])
-                self.fix_step.params["rem"]["occupations"] = 2
-                self.fix_step.params["rem"]["fon_norb"] = int(natoms * 0.618)
-                self.fix_step.params["rem"]["fon_t_start"] = 300
-                self.fix_step.params["rem"]["fon_t_end"] = 300
-                self.fix_step.params["rem"]["fon_e_thresh"] = 6
-                self.fix_step.set_integral_threshold(14)
-                self.fix_step.set_scf_convergence_threshold(7)
-            else:
-                raise ValueError("fix method " + method + " is not supported")
-            strategy_text = "<SCF Fix Strategy>"
-            strategy_text += json.dumps(strategy, indent=4, sort_keys=True)
-            strategy_text += "</SCF Fix Strategy>"
-            if len(old_strategy_text) > 0:
-                comments = scf_pattern.sub(strategy_text, comments)
-            else:
-                comments += "\n" + strategy_text
-            self.fix_step.params["comment"] = comments
-            return method
-
-    def set_last_input_geom(self, new_mol):
-        for i in range(self.error_step_id, -1, -1):
-            qctask = self.qcinp.jobs[i]
-            if isinstance(qctask.mol, Molecule):
-                qctask.mol = copy.deepcopy(new_mol)
-
-    def set_scf_initial_guess(self, guess="sad"):
-        if "scf_guess" not in self.fix_step.params["rem"] \
-                or self.error_step_id > 0 \
-                or self.fix_step.params["rem"]["scf_guess"] != "read":
-            self.fix_step.set_scf_initial_guess(guess)
-
-    def fix_geom_opt(self):
-        comments = self.fix_step.params.get("comment", "")
-        geom_pattern = re.compile(r"<Geom Opt Fix Strategy>(.*)"
-                                  r"</Geom Opt Fix Strategy>",
-                                  flags=re.DOTALL)
-        old_strategy_text = re.findall(geom_pattern, comments)
-
-        if len(old_strategy_text) > 0:
-            old_strategy_text = old_strategy_text[0]
-
-        od = self.outdata[self.error_step_id]
-
-        if 'Lamda Determination Failed' in self.errors and len(od["molecules"])>=2:
-            self.fix_step.set_scf_algorithm_and_iterations(
-                algorithm="diis", iterations=self.scf_max_cycles)
-            if self.error_step_id > 0:
-                self.set_scf_initial_guess("read")
-            else:
-                self.set_scf_initial_guess("sad")
-            self.set_last_input_geom(od["molecules"][-1])
-            if od["jobtype"] == "aimd":
-                aimd_steps = self.fix_step.params["rem"]["aimd_steps"]
-                elapsed_steps = len(od["molecules"]) - 1
-                remaining_steps = aimd_steps - elapsed_steps + 1
-                self.fix_step.params["rem"]["aimd_steps"] = remaining_steps
-            if len(old_strategy_text) > 0:
-                comments = geom_pattern.sub("", comments)
-                self.fix_step.params["comment"] = comments
-                if len(comments.strip()) == 0:
-                    self.fix_step.params.pop("comment")
-            return "reset"
-
-        if len(od["molecules"]) <= 10:
-            # immature termination of geometry optimization
-            if 'Exit Code 134' in self.errors:
-                return self.fix_error_code_134()
-            else:
-                return None
-        if len(old_strategy_text) > 0:
-            strategy = json.loads(old_strategy_text)
-            strategy["current_method_id"] += 1
-        else:
-            strategy = dict()
-            strategy["methods"] = ["increase_iter", "GDIIS", "CartCoords"]
-            strategy["current_method_id"] = 0
-        if strategy["current_method_id"] > len(strategy["methods"]) - 1:
-            return None
-        else:
-            method = strategy["methods"][strategy["current_method_id"]]
-            if method == "increase_iter":
-                self.fix_step.set_geom_max_iterations(self.geom_max_cycles)
-                self.set_last_input_geom(od["molecules"][-1])
-            elif method == "GDIIS":
-                self.fix_step.set_geom_opt_use_gdiis(subspace_size=5)
-                self.fix_step.set_geom_max_iterations(self.geom_max_cycles)
-                self.set_last_input_geom(od["molecules"][-1])
-            elif method == "CartCoords":
-                self.fix_step.set_geom_opt_coords_type("cartesian")
-                self.fix_step.set_geom_max_iterations(self.geom_max_cycles)
-                self.fix_step.set_geom_opt_use_gdiis(0)
-                self.set_last_input_geom(od["molecules"][-1])
-            else:
-                raise ValueError("fix method" + method + "is not supported")
-            strategy_text = "<Geom Opt Fix Strategy>"
-            strategy_text += json.dumps(strategy, indent=4, sort_keys=True)
-            strategy_text += "</Geom Opt Fix Strategy>"
-            if len(old_strategy_text) > 0:
-                comments = geom_pattern.sub(strategy_text, comments)
-            else:
-                comments += "\n" + strategy_text
-            self.fix_step.params["comment"] = comments
-            return method
-
-    def backup(self):
-        error_num = max([0] + [int(f.split(".")[1])
-                               for f in glob.glob("error.*.tar.gz")])
-        filename = "error.{}.tar.gz".format(error_num + 1)
-        logging.info("Backing up run to {}.".format(filename))
-        tar = tarfile.open(filename, "w:gz")
-        bak_list = [self.input_file, self.output_file] + \
-            list(self.ex_backup_list)
-        for f in bak_list:
-            if os.path.exists(f):
-                tar.add(f)
-        tar.close()
-
-    def as_dict(self):
-        return {"@module": self.__class__.__module__,
-                "@class": self.__class__.__name__,
-                "input_file": self.input_file,
-                "output_file": self.output_file,
-                "ex_backup_list": tuple(self.ex_backup_list),
-                "rca_gdm_thresh": self.rca_gdm_thresh,
-                "scf_max_cycles": self.scf_max_cycles,
-                "geom_max_cycles": self.geom_max_cycles,
-                "outdata": self.outdata,
-                "qcinp": self.qcinp.as_dict() if self.qcinp else None,
-                "error_step_id": self.error_step_id,
-                "errors": self.errors,
-                "fix_step": self.fix_step.as_dict() if self.fix_step else None}
-
-    @classmethod
-    def from_dict(cls, d):
-        h = QChemErrorHandler(input_file=d["input_file"],
-                              output_file=d["output_file"],
-                              ex_backup_list=d["ex_backup_list"],
-                              rca_gdm_thresh=d["rca_gdm_thresh"],
-                              scf_max_cycles=d["scf_max_cycles"],
-                              geom_max_cycles=d["geom_max_cycles"])
-        h.outdata = d["outdata"]
-        h.qcinp = QcInput.from_dict(d["qcinp"]) if d["qcinp"] else None
-        h.error_step_id = d["error_step_id"]
-        h.errors = d["errors"]
-        h.fix_step = QcTask.from_dict(d["fix_step"]) if d["fix_step"] else None
-        return h
+        print("This hasn't been implemented yet!")
+        return {"errors": self.errors, "actions": None}
